Issue 2715: Merged existing changes with new commits to master.
