+++ test/system/src/test/java/io/pravega/test/system/MultiControllerTest.java
@@ -43,14 +43,13 @@ import static io.pravega.test.system.framework.Utils.DOCKER_BASED;
 @Slf4j
 @RunWith(SystemTestRunner.class)
 public class MultiControllerTest {
-    private static final ScheduledExecutorService EXECUTOR_SERVICE = Executors.newSingleThreadScheduledExecutor();
+    private final ScheduledExecutorService executorService = Executors.newSingleThreadScheduledExecutor();
     private Service controllerService = null;
     private AtomicReference<URI> controllerURIDirect = new AtomicReference<>();
     private AtomicReference<URI> controllerURIDiscover = new AtomicReference<>();
 
-
     @Environment
-    public static void setup() throws MarathonException, ExecutionException {
+    public static void initialize() throws MarathonException, ExecutionException {
         Service zkService = Utils.createZookeeperService();
         if (!zkService.isRunning()) {
             zkService.start(true);
@@ -66,7 +65,6 @@ public class MultiControllerTest {
 
         List<URI> conUris = controllerService.getServiceDetails();
         log.debug("Pravega Controller service  details: {}", conUris);
-
     }
 
     @Before
@@ -96,12 +94,11 @@ public class MultiControllerTest {
 
     @After
     public void tearDown() {
-        ExecutorServiceHelpers.shutdown(EXECUTOR_SERVICE);
+        ExecutorServiceHelpers.shutdown(executorService);
         if (controllerService != null && controllerService.isRunning()) {
             controllerService.stop();
             controllerService.clean();
         }
-
     }
 
     /**
@@ -163,7 +160,6 @@ public class MultiControllerTest {
             Assert.assertTrue(createScopeWithSimpleRetry(
                     "scope" + RandomStringUtils.randomAlphanumeric(10), controllerURIDiscover.get()));
         }
-
     }
 
     private boolean createScopeWithSimpleRetry(String scopeName, URI controllerURI) throws ExecutionException, InterruptedException {
@@ -173,12 +169,12 @@ public class MultiControllerTest {
                 .clientConfig(ClientConfig.builder()
                         .controllerURI(controllerURI)
                         .build())
-                .build(), EXECUTOR_SERVICE);
+                .build(), executorService);
 
         CompletableFuture<Boolean> retryResult = Retry.withExpBackoff(500, 2, 10, 5000)
                 .retryingOn(Exception.class)
                 .throwingOn(IllegalArgumentException.class)
-                .runAsync(() -> controllerClient.createScope(scopeName), EXECUTOR_SERVICE);
+                .runAsync(() -> controllerClient.createScope(scopeName), executorService);
 
         return retryResult.get();
     }
@@ -189,7 +185,7 @@ public class MultiControllerTest {
                 .clientConfig(ClientConfig.builder()
                         .controllerURI(controllerURI)
                         .build())
-                .build(), EXECUTOR_SERVICE);
+                .build(), executorService);
         return controllerClient.createScope(scopeName).get();
     }
 }
