+++ test/system/src/test/java/io/pravega/test/system/PravegaSegmentStoreTest.java
@@ -31,7 +31,7 @@ public class PravegaSegmentStoreTest {
      * @throws MarathonException if error in setup
      */
     @Environment
-    public static void setup() throws MarathonException {
+    public static void initialize() throws MarathonException {
         Service zk = Utils.createZookeeperService();
         if (!zk.isRunning()) {
             zk.start(true);
