+++ test/system/src/test/java/io/pravega/test/system/ZookeeperTest.java
@@ -33,7 +33,7 @@ public class ZookeeperTest {
      * @throws MarathonException if error in setup
      */
     @Environment
-    public static void setup() throws MarathonException {
+    public static void initialize() throws MarathonException {
         Service zk = Utils.createZookeeperService();
         if (!zk.isRunning()) {
             zk.start(true);
